Bug fixed & Structure modified
