+++ package.json
@@ -7,6 +7,7 @@
     "@babel/core": "7.12.3",
     "@pmmmwh/react-refresh-webpack-plugin": "0.4.3",
     "@svgr/webpack": "5.5.0",
+    "@systemop/localstorage-fs": "1.0.0",
     "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
@@ -88,8 +89,7 @@
     "webpack": "4.44.2",
     "webpack-dev-server": "3.11.1",
     "webpack-manifest-plugin": "2.2.0",
-    "workbox-webpack-plugin": "5.1.4",
-    "@systemop/localstorage-fs": "1.0.0"
+    "workbox-webpack-plugin": "5.1.4"
   },
   "scripts": {
     "start": "concurrently \"node scripts/start.js\" \"nodemon ./server/server.js\"",
