+++ src/components/BreadCrumb/index.js
@@ -1,5 +1,14 @@
-import BreadCrumbContainer from "./BreadCrumbContainer";
-
-const BreadCrumb = BreadCrumbContainer;
+import BreadCrumbPresenter from "./BreadCrumbPresenter"
+import React from "react"
+import {get_root_pointer} from "@systemop/localstorage-fs"
+const BreadCrumb = (cwd) => {
+    const root_pointer = get_root_pointer() ;
+    
+    return(
+        <BreadCrumbPresenter
+            cwd ={cwd}>
+        </BreadCrumbPresenter>
+    );
+};
 
 export default BreadCrumb;
