+++ src/components/List/index.js
@@ -1,4 +1,17 @@
-import ListContainer from "./ListContainer";
+import ListPresenter from "./ListPresenter";
+import React from "react";
+import { observer, inject } from "mobx-react"
 
+const List = ({
+  showCheckbox,
+  checkedItemHandler,
+}) => {
+  return (
+    <ListPresenter
+      showCheckbox={showCheckbox}
+      checkedItemHandler={checkedItemHandler}
+    />
+  );
+};
 
-export default ListContainer;
+export default inject("storeEditor", "storeData")(observer(List));
