+++ src/components/ListDirItem/index.js
@@ -1,5 +1,36 @@
-import ListDirItemContainer from "./ListDirItemContainer";
+import ListDirItemPresenter from "./ListDirItemPresenter";
+import { React, useState } from "react";
+import dateFormat, { masks } from "dateformat";
+import { observer, inject } from "mobx-react"
 
-const ListDirItem = ListDirItemContainer;
+const ListDirItem = ({
+  storeData,
+  name,
+  time,
+  set
+}) => {
+  const dataManage = storeData;
+  let timeString = (time != null)? (dateFormat(time, "yyyy. m. d  HH:MM")) : null;
+  const onChange = (name) => {
+    if(set === "parent"){
+      dataManage.setParentDir(name);
+      dataManage.setDirList();
+      dataManage.setDataList();
+    } else{
+      dataManage.setChangeDir(name);
+      dataManage.setDirList();
+      dataManage.setDataList();
+    }
+  }
+  return (
+    <div onClick = {() => onChange(name)}>
+      <ListDirItemPresenter
+        name={name}
+        time={timeString}
+      />
+    </div>
+  );
+};
 
-export default ListDirItem;
+export default inject("storeData")(observer(ListDirItem));
+
