--- src/components/MainApp/HeaderContainer.js
@@ -1,7 +1,0 @@
-import HeaderPresenter from "./HeaderPresenter";
-
-const HeaderContainer = () => {
-  return <HeaderPresenter />;
-};
-
-export default HeaderContainer;
