--- src/components/PostIt/PostItContainer.js
@@ -1,17 +1,0 @@
-import PostItPresenter from "./PostItPresenter";
-import React from "react";
-import { observer, inject } from "mobx-react"
-
-const PostItContainer = ({
-  showCheckbox,
-  checkedItemHandler,
-}) => {
-  return (
-    <PostItPresenter
-      showCheckbox={showCheckbox}
-      checkedItemHandler={checkedItemHandler}
-    />
-  );
-};
-
-export default inject("storeEditor", "storeData")(observer(PostItContainer));
