+++ src/components/PostIt/index.js
@@ -1,5 +1,17 @@
-import PostItContainer from "./PostItContainer";
+import PostItPresenter from "./PostItPresenter";
+import React from "react";
+import { observer, inject } from "mobx-react"
 
-const PostIt = PostItContainer;
+const PostIt = ({
+  showCheckbox,
+  checkedItemHandler,
+}) => {
+  return (
+    <PostItPresenter
+      showCheckbox={showCheckbox}
+      checkedItemHandler={checkedItemHandler}
+    />
+  );
+};
 
-export default PostIt;
+export default inject("storeEditor", "storeData")(observer(PostIt));
