--- src/components/PostItDirItem/PostItDirItemContainer.js
@@ -1,28 +1,0 @@
-import PostItDirItemPresenter from "./PostItDirItemPresenter";
-import { observer, inject } from "mobx-react"
-
-const PostItDirItemContainer = ({
-  storeData,
-  name,
-  set
-}) => {
-  const dataManage = storeData;
-  const onChange = (name) => {
-    if(set === "parent"){
-      dataManage.setParentDir(name);
-    } else{
-      dataManage.setChangeDir(name);
-    }
-    dataManage.setDirList();
-    dataManage.setDataList();
-  }
-  return (
-    <div onClick={() => onChange(name)}>
-      <PostItDirItemPresenter
-        name={name}
-      />
-    </div>
-  );
-};
-
-export default inject("storeData")(observer(PostItDirItemContainer));
