+++ src/components/PostItDirItem/index.js
@@ -1,5 +1,28 @@
-import PostItDirItemContainer from "./PostItDirItemContainer";
+import PostItDirItemPresenter from "./PostItDirItemPresenter";
+import { observer, inject } from "mobx-react"
 
-const PostItDirItem = PostItDirItemContainer;
+const PostItDirItem = ({
+  storeData,
+  name,
+  set
+}) => {
+  const dataManage = storeData;
+  const onChange = (name) => {
+    if(set === "parent"){
+      dataManage.setParentDir(name);
+    } else{
+      dataManage.setChangeDir(name);
+    }
+    dataManage.setDirList();
+    dataManage.setDataList();
+  }
+  return (
+    <div onClick={() => onChange(name)}>
+      <PostItDirItemPresenter
+        name={name}
+      />
+    </div>
+  );
+};
 
-export default PostItDirItem;
+export default inject("storeData")(observer(PostItDirItem));
