+++ src/components/PostItItem/index.js
@@ -1,6 +1,34 @@
-import PostItDirItemContainer from "../PostItDirItem/PostItDirItemContainer";
-import PostItItemContainer from "./PostItItemContainer";
+import PostItItemPresenter from "./PostItItemPresenter";
+import { useState } from "react";
+import { observer, inject } from "mobx-react"
 
-const PostItItem = PostItItemContainer;
+const PostItItem = ({
+  storeEditor,
+  title,
+  content,
+  uid,
+  showCheckbox,
+  checkedItemHandler,
+}) => {
+  const [bChecked, setChecked] = useState(false);
+  const controlEditor = storeEditor;
 
-export default PostItItem;
+  const checkHandler = ({ target }) => {
+    setChecked(!bChecked);
+    checkedItemHandler(uid, target.checked); //reflect change on del item list
+  };
+
+  return (
+    <div onClick = {() => controlEditor.setId(uid)}>
+      <PostItItemPresenter
+        title={title}
+        content={content}
+        showCheckbox={showCheckbox}
+        bChecked={bChecked}
+        checkHandler={checkHandler}
+      />
+    </div>
+  );
+};
+
+export default inject("storeEditor")(observer(PostItItem));
